requirements.txt
git+https://github.com/openai/CLIP.git
torch
torchvision
tqdm
Pillow
numpy
matplotlib
